split file into two
